refactor(web): move getLatestSafeVersion to @safe-global/utils
