+++ apps/web/src/components/new-safe/create/index.tsx
@@ -19,9 +19,9 @@ import { type ReactElement, useMemo, useState } from 'react'
 import ExternalLink from '@/components/common/ExternalLink'
 import { HelpCenterArticle } from '@/config/constants'
 import { type SafeVersion } from '@safe-global/safe-core-sdk-types'
-import { getLatestSafeVersion } from '@/utils/chains'
 import { useCurrentChain } from '@/hooks/useChains'
 import type { ChainInfo } from '@safe-global/safe-gateway-typescript-sdk'
+import { getLatestSafeVersion } from '@safe-global/utils/utils/chains'
 
 export type NewSafeFormData = {
   name: string
