+++ apps/web/src/components/new-safe/create/steps/SetNameStep/index.tsx
@@ -16,7 +16,6 @@ import NoWalletConnectedWarning from '../../NoWalletConnectedWarning'
 import { type SafeVersion } from '@safe-global/safe-core-sdk-types'
 import { useCurrentChain } from '@/hooks/useChains'
 import { useEffect } from 'react'
-import { getLatestSafeVersion } from '@/utils/chains'
 import type { ChainInfo } from '@safe-global/safe-gateway-typescript-sdk'
 import { useSafeSetupHints } from '../OwnerPolicyStep/useSafeSetupHints'
 import type { CreateSafeInfoItem } from '../../CreateSafeInfos'
@@ -24,6 +23,7 @@ import NetworkMultiSelector from '@/components/common/NetworkSelector/NetworkMul
 import { useAppSelector } from '@/store'
 import { selectChainById } from '@/store/chainsSlice'
 import useWallet from '@/hooks/wallets/useWallet'
+import { getLatestSafeVersion } from '@safe-global/utils/utils/chains'
 
 type SetNameStepForm = {
   name: string
