+++ apps/web/src/features/recovery/services/__tests__/recovery-state.test.ts
@@ -15,8 +15,8 @@ import { useWeb3ReadOnly } from '@/hooks/wallets/web3'
 import { encodeMultiSendData } from '@safe-global/protocol-kit/dist/src/utils/transactions/utils'
 import { getMultiSendCallOnlyDeployment, getSafeSingletonDeployment } from '@safe-global/safe-deployments'
 import { Interface } from 'ethers'
-import { getLatestSafeVersion } from '@/utils/chains'
 import { chainBuilder } from '@/tests/builders/chains'
+import { getLatestSafeVersion } from '@safe-global/utils/utils/chains'
 
 jest.mock('@/hooks/wallets/web3')
