+++ apps/web/src/services/tx/tx-sender/dispatch.ts
@@ -36,7 +36,8 @@ import { createWeb3, getUserNonce } from '@/hooks/wallets/web3'
 import { asError } from '@/services/exceptions/utils'
 import chains from '@/config/chains'
 import { createExistingTx } from './create'
-import { getLatestSafeVersion } from '@/utils/chains'
+
+import { getLatestSafeVersion } from '@safe-global/utils/utils/chains'
 
 /**
  * Propose a transaction
