+++ apps/web/src/utils/__tests__/chains.test.ts
@@ -1,4 +1,5 @@
-import { hasFeature, getBlockExplorerLink, FEATURES, getLatestSafeVersion } from '@/utils/chains'
+import { hasFeature, getBlockExplorerLink, FEATURES } from '@/utils/chains'
+import { getLatestSafeVersion } from '@safe-global/utils/utils/chains'
 import { CONFIG_SERVICE_CHAINS } from '@/tests/mocks/chains'
 import { chainBuilder } from '@/tests/builders/chains'
